Removing peerDependencies
While a good idea in theory, they are causing way too much trouble with
generator-bbb.  Currently since generator-bbb is listed as a peerDep to
Backbone Boilerplate and Backbone Boilerplate is a dependency of
generator-bbb, we have a circular dependency which causes problems when
you fix a bug in generator-bbb, but backbone boilerplate pulls the wrong
version and still breaks.