Added Views and removed the styles Grunt plugin
The styles Grunt plugin was really cool, but I don't have time to
maintain it. Since it never left beta, I'm going to remove it in favor
of the Grunt contrib team's cssmin. If you want to use a preprocessor,
it's trivial to add.

I've added a bunch of application files to demonstrate how to integrate
the router, components, etc. into your application.